+++ Read-and-Save-document/Multithreading-using-parallel-process/.NET/Multithreading-using-parallel-process/Multithreading-using-parallel-process.csproj
@@ -3,11 +3,11 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>Multithreaded_using_parallel_process</RootNamespace>
+    <RootNamespace>Multithreading_using_parallel_process</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Syncfusion.DocIORenderer.Net.Core" Version="*" />
+    <PackageReference Include="Syncfusion.DocIO.Net.Core" Version="*" />
   </ItemGroup>
 
   <ItemGroup>
