+++ Read-and-Save-document/Multithreading-using-parallel-process/.NET/Multithreading-using-parallel-process/Program.cs
@@ -4,26 +4,25 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 
-namespace Multithreaded_using_tasks
+namespace Multithreading_using_parallel_process
 {
     class MultiThreading
     {
-        //Indicates the number of threads to be create.
-        private const int TaskCount = 1000;
-        public static async Task Main()
+        static void Main(string[] args)
         {
-            //Create an array of tasks based on the TaskCount.
-            Task[] tasks = new Task[TaskCount];
-            for (int i = 0; i < TaskCount; i++)
+            //Indicates the number of threads to be create.
+            int limit = 5;
+            Console.WriteLine("Parallel For Loop");
+            Parallel.For(0, limit, count =>
             {
-                tasks[i] = Task.Run(() => OpenAndSaveWordDocument());
-            }
-            //Ensure all tasks complete by waiting on each task.
-            await Task.WhenAll(tasks);
+                Console.WriteLine("Task {0} started", count);
+                //Create multiple Word document, one document on each thread.
+                OpenAndSaveWordDocument(count);
+                Console.WriteLine("Task {0} is done", count);
+            });
         }
-
         //Open and save a Word document using multi-threading.
-        static void OpenAndSaveWordDocument()
+        static void OpenAndSaveWordDocument(int count)
         {
             using (FileStream inputStream = new FileStream(Path.GetFullPath(@"Data/Input.docx"), FileMode.Open, FileAccess.Read))
             {
@@ -33,7 +32,7 @@ namespace Multithreaded_using_tasks
                     //Add text to the last paragraph.
                     document.LastParagraph.AppendText("Product Overview");
                     //Save the Word document in the desired format.
-                    using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + Guid.NewGuid().ToString() + ".docx"), FileMode.Create, FileAccess.Write))
+                    using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + count + ".docx"), FileMode.Create, FileAccess.Write))
                     {
                         document.Save(outputFileStream, FormatType.Docx);
                     }
