+++ Read-and-Save-document/Multithreading-using-tasks/.NET/Multithreading-using-tasks/Program.cs
@@ -4,25 +4,26 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 
-namespace Multithreaded_using_parallel_process
+namespace Multithreading_using_tasks
 {
     class MultiThreading
     {
-        static void Main(string[] args)
+        //Indicates the number of threads to be create.
+        private const int TaskCount = 1000;
+        public static async Task Main()
         {
-            //Indicates the number of threads to be create.
-            int limit = 5;
-            Console.WriteLine("Parallel For Loop");
-            Parallel.For(0, limit, count =>
+            //Create an array of tasks based on the TaskCount.
+            Task[] tasks = new Task[TaskCount];
+            for (int i = 0; i < TaskCount; i++)
             {
-                Console.WriteLine("Task {0} started", count);
-                //Create multiple Word document, one document on each thread.
-                OpenAndSaveWordDocument(count);
-                Console.WriteLine("Task {0} is done", count);
-            });
+                tasks[i] = Task.Run(() => OpenAndSaveWordDocument());
+            }
+            //Ensure all tasks complete by waiting on each task.
+            await Task.WhenAll(tasks);
         }
+
         //Open and save a Word document using multi-threading.
-        static void OpenAndSaveWordDocument(int count)
+        static void OpenAndSaveWordDocument()
         {
             using (FileStream inputStream = new FileStream(Path.GetFullPath(@"Data/Input.docx"), FileMode.Open, FileAccess.Read))
             {
@@ -32,7 +33,7 @@ namespace Multithreaded_using_parallel_process
                     //Add text to the last paragraph.
                     document.LastParagraph.AppendText("Product Overview");
                     //Save the Word document in the desired format.
-                    using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + count + ".docx"), FileMode.Create, FileAccess.Write))
+                    using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + Guid.NewGuid().ToString() + ".docx"), FileMode.Create, FileAccess.Write))
                     {
                         document.Save(outputFileStream, FormatType.Docx);
                     }
