+++ Word-to-PDF-Conversion/Multithreading-using-parallel-process/.NET/Multithreading-using-parallel-process/Multithreading-using-parallel-process.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>Multithreaded_using_tasks</RootNamespace>
+    <RootNamespace>Multithreading_using_parallel_process</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
