+++ Word-to-PDF-Conversion/Multithreading-using-parallel-process/.NET/Multithreading-using-parallel-process/Program.cs
@@ -6,26 +6,25 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 
-namespace Multithreaded_using_tasks
+namespace Multithreading_using_parallel_process
 {
     class MultiThreading
     {
-        //Indicates the number of threads to be create.
-        private const int TaskCount = 1000;
-        public static async Task Main()
+        static void Main(string[] args)
         {
-            //Create an array of tasks based on the TaskCount.
-            Task[] tasks = new Task[TaskCount];
-            for (int i = 0; i < TaskCount; i++)
+            //Indicates the number of threads to be create.
+            int limit = 5;
+            Console.WriteLine("Parallel For Loop");
+            Parallel.For(0, limit, count =>
             {
-                tasks[i] = Task.Run(() => ConvertWordToPDF());
-            }
-            //Ensure all tasks complete by waiting on each task.
-            await Task.WhenAll(tasks);
+                Console.WriteLine("Task {0} started", count);
+                //Convert multiple Word document, one document on each thread.
+                ConvertWordToPDF(count);
+                Console.WriteLine("Task {0} is done", count);
+            });
         }
-
         //Convert a Word document to PDF using multi-threading.
-        static void ConvertWordToPDF()
+        static void ConvertWordToPDF(int count)
         {
             using (FileStream inputStream = new FileStream(Path.GetFullPath(@"Data/Input.docx"), FileMode.Open, FileAccess.Read))
             {
@@ -39,7 +38,7 @@ namespace Multithreaded_using_tasks
                         using (PdfDocument pdfDocument = renderer.ConvertToPDF(document))
                         {
                             //Save the PDF document.
-                            using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + Guid.NewGuid().ToString() + ".pdf"), FileMode.Create, FileAccess.Write))
+                            using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + count + ".pdf"), FileMode.Create, FileAccess.Write))
                             {
                                 pdfDocument.Save(outputFileStream);
                             }
