+++ Word-to-PDF-Conversion/Multithreading-using-tasks/.NET/Multithreading-using-tasks.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31911.196
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Multithreaded-using-parallel-process", "Multithreaded-using-parallel-process\Multithreaded-using-parallel-process.csproj", "{C17B90BC-F559-456B-B189-90B53FF6CDD4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Multithreading-using-tasks", "Multithreading-using-tasks\Multithreading-using-tasks.csproj", "{C17B90BC-F559-456B-B189-90B53FF6CDD4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
