+++ Word-to-PDF-Conversion/Multithreading-using-tasks/.NET/Multithreading-using-tasks/Multithreading-using-tasks.csproj
@@ -3,11 +3,11 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>Multithreaded_using_parallel_process</RootNamespace>
+    <RootNamespace>Multithreading_using_tasks</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Syncfusion.DocIO.Net.Core" Version="*" />
+    <PackageReference Include="Syncfusion.DocIORenderer.Net.Core" Version="*" />
   </ItemGroup>
 
   <ItemGroup>
